Try DELETING SQS messages, even on fail.
Most of our errors aren't fixed by retrying. They tend to be missing key
errors, or deeper bugs that require code changes. All retrying does
is spike CPU and back up the queue.

When possible, we will try to delete SQS messages, even in the catch. We will
be sure to log errors and monitor the logs, but bogging down message queues
helps no one.

A common pattern in TT is for a user to upload a few hundred photos all at
once and then go through and delete the duplicates, etc. As the virus scanner
catches up, this can lead to 404s and "No such key errors". We will never
be able to recover from these, since the key is _not_ coming back. So we will
simply delete the SQS message and move on.

A future enhancement could be to report errors on an SNS topic to be consumed
somewhere else (possibly lambda).